+++ icevision/datasets/coco/__init__.py
@@ -1,2 +1,0 @@
-from icevision.datasets.coco.data import *
-from icevision.datasets.coco.parsers import *