+++ icevision/datasets/fridge/__init__.py
@@ -1,3 +1,0 @@
-from icevision.datasets.fridge.data import *
-from icevision.datasets.fridge.parsers import *
-from icevision.datasets.fridge.visualization import *