+++ icevision/datasets/pennfundan/__init__.py
@@ -1,2 +1,0 @@
-from icevision.datasets.pennfundan.data import *
-from icevision.datasets.pennfundan.parsers import *