+++ tests/datasets/coco/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture()
-def coco_annotations_file(samples_source):
-    return samples_source / "annotations.json"