+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed package from `mantisshrimp` to `icevision`
 
 ## Deleted
+- Removed `datasets` module to instead use the new `icedata` package
 
 ## [0.0.9]